feat: rename `scheduler/resource` to `scheduler/resource/standard`
