+++ scheduler/networktopology/probes.go
@@ -31,7 +31,7 @@ import (
 	"d7y.io/dragonfly/v2/pkg/cache"
 	pkgredis "d7y.io/dragonfly/v2/pkg/redis"
 	"d7y.io/dragonfly/v2/scheduler/config"
-	"d7y.io/dragonfly/v2/scheduler/resource"
+	resource "d7y.io/dragonfly/v2/scheduler/resource/standard"
 )
 
 const (
