+++ scheduler/resource/standard/host_manager.go
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-//go:generate mockgen -destination host_manager_mock.go -source host_manager.go -package resource
+//go:generate mockgen -destination host_manager_mock.go -source host_manager.go -package standard
 
-package resource
+package standard
 
 import (
 	"sync"
