+++ scheduler/resource/standard/host_manager_mock.go
@@ -3,11 +3,11 @@
 //
 // Generated by this command:
 //
-//	mockgen -destination host_manager_mock.go -source host_manager.go -package resource
+//	mockgen -destination host_manager_mock.go -source host_manager.go -package standard
 //
 
-// Package resource is a generated GoMock package.
-package resource
+// Package standard is a generated GoMock package.
+package standard
 
 import (
 	reflect "reflect"
