+++ scheduler/resource/standard/peer_manager.go
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-//go:generate mockgen -destination peer_manager_mock.go -source peer_manager.go -package resource
+//go:generate mockgen -destination peer_manager_mock.go -source peer_manager.go -package standard
 
-package resource
+package standard
 
 import (
 	"context"
