+++ scheduler/resource/standard/peer_manager_mock.go
@@ -3,11 +3,11 @@
 //
 // Generated by this command:
 //
-//	mockgen -destination peer_manager_mock.go -source peer_manager.go -package resource
+//	mockgen -destination peer_manager_mock.go -source peer_manager.go -package standard
 //
 
-// Package resource is a generated GoMock package.
-package resource
+// Package standard is a generated GoMock package.
+package standard
 
 import (
 	reflect "reflect"
