+++ scheduler/resource/standard/resource_mock.go
@@ -3,11 +3,11 @@
 //
 // Generated by this command:
 //
-//	mockgen -destination resource_mock.go -source resource.go -package resource
+//	mockgen -destination resource_mock.go -source resource.go -package standard
 //
 
-// Package resource is a generated GoMock package.
-package resource
+// Package standard is a generated GoMock package.
+package standard
 
 import (
 	reflect "reflect"
