+++ scheduler/resource/standard/seed_peer_client_mock.go
@@ -3,11 +3,11 @@
 //
 // Generated by this command:
 //
-//	mockgen -destination seed_peer_client_mock.go -source seed_peer_client.go -package resource
+//	mockgen -destination seed_peer_client_mock.go -source seed_peer_client.go -package standard
 //
 
-// Package resource is a generated GoMock package.
-package resource
+// Package standard is a generated GoMock package.
+package standard
 
 import (
 	context "context"
