+++ scheduler/resource/standard/seed_peer_test.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package resource
+package standard
 
 import (
 	"context"
