+++ scheduler/resource/standard/task_manager_mock.go
@@ -3,11 +3,11 @@
 //
 // Generated by this command:
 //
-//	mockgen -destination task_manager_mock.go -source task_manager.go -package resource
+//	mockgen -destination task_manager_mock.go -source task_manager.go -package standard
 //
 
-// Package resource is a generated GoMock package.
-package resource
+// Package standard is a generated GoMock package.
+package standard
 
 import (
 	reflect "reflect"
